+++ api/commands/chia_cli.py
@@ -117,7 +117,7 @@ def load_blockchain_show(blockchain):
 
 def load_connections_show(blockchain):
     chia_binary = globals.get_blockchain_binary(blockchain)
-    if blockchain == 'cactus':  # Cactus now supports only 'peer' command
+    if blockchain in ['cactus', 'littlelambocoin', 'maize']:  # These now support only the 'peer' command
         proc = Popen("{0} peer -c full_node".format(chia_binary), stdout=PIPE, stderr=PIPE, shell=True)
     else:
         proc = Popen("{0} show --connections".format(chia_binary), stdout=PIPE, stderr=PIPE, shell=True)
@@ -208,7 +208,7 @@ def pause_wallet(blockchain):
 def remove_connection(node_id, ip, blockchain):
     chia_binary = globals.get_blockchain_binary(blockchain)
     try:
-        if blockchain == 'cactus':  # Cactus now supports only 'peer' command
+        if blockchain in ['cactus', 'littlelambocoin', 'maize']:  # These now support only the 'peer' command
             proc = Popen("{0} peer --remove-connection {1} full_node".format(chia_binary, node_id), stdout=PIPE, stderr=PIPE, shell=True)
         else:
             proc = Popen("{0} show --remove-connection {1}".format(chia_binary, node_id), stdout=PIPE, stderr=PIPE, shell=True)
@@ -254,7 +254,7 @@ def plot_check(blockchain, plot_path):
         raise Exception("The timeout is expired attempting to check plots.")
     app.logger.info("Completed plot check of: {0}".format(plot_path))
     class_escape = re.compile(r'.*: INFO\s+')
-    ansi_escape = re.compile(r'\x1B(?:[@-Z\\-_]|\[[0-?]*[ -/]*[@-~])')
+    ansi_escape = re.compile(r'\x1B(?:[@A-Z\\-_]|\[[0-9:;<=>?]*[ -/]*[@-~])')
     return  class_escape.sub('', ansi_escape.sub('', outs))
 
 def dispatch_action(job):
@@ -292,7 +292,7 @@ def add_connections(connections, blockchain):
             elif socket.gethostbyname(hostname) != hostname:
                 app.logger.debug('{} is a valid hostname'.format(hostname))
             app.logger.info("Adding {0} connection to peer: {1}".format(blockchain, connection))
-            if blockchain == 'cactus':  # Cactus now supports only 'peer' command
+            if blockchain in ['cactus', 'littlelambocoin', 'maize']:  # These now support only the 'peer' command
                 proc = Popen("{0} peer --add-connection {1} full_node".format(chia_binary, connection), stdout=PIPE, stderr=PIPE, shell=True)
             else:
                 proc = Popen("{0} show --add-connection {1}".format(chia_binary, connection), stdout=PIPE, stderr=PIPE, shell=True)
@@ -333,11 +333,17 @@ def remove_connection(node_ids, blockchain):
 
 def save_wallet_settings(settings, blockchain):
     try:
+        app.logger.info("Setting wallet frequency: {0}".format(settings))
         if not settings: # User reverting to defaults, no custom settings
-            os.path(WALLET_SETTINGS_FILE).delete()
+            app.logger.info("Deleting settings at {0}".format(WALLET_SETTINGS_FILE))
+            try:
+                os.remove(WALLET_SETTINGS_FILE)
+            except OSError:
+                pass
         else:
+            app.logger.info("Updating settings at {0}".format(WALLET_SETTINGS_FILE))
             with open(WALLET_SETTINGS_FILE, 'w') as outfile:
                 json.dump(settings, outfile)
     except Exception as ex:
-        app.logger.info(traceback.format_exc())
+        app.logger.debug(traceback.format_exc())
         raise Exception('Failed to store {0} wallet settings to {1}.'.format(blockchain, WALLET_SETTINGS_FILE) + '\n' + str(ex))