+++ api/commands/fd_cli.py
@@ -20,7 +20,7 @@ def reward_recovery(wallet_id, launcher_id, pool_contract_address):
     if not rpc_port:
         app.logger.info("Skipping NFT reward recovery on unsupported blockchain: {0}".format(blockchain))
         return
-    logfile = "/root/.chia/machinaris/logs/fd-cli.log"
+    logfile = "/root/.chia/machinaris/logs/rewards.log"
     log_fd = os.open(logfile, os.O_RDWR | os.O_CREAT)
     log_fo = os.fdopen(log_fd, "a+")
     vars = {}
@@ -49,7 +49,7 @@ def reward_recovery(wallet_id, launcher_id, pool_contract_address):
     fd_env.update(vars)
     cmd = f"/usr/local/bin/fd-cli nft-recover -l {launcher_id} -p {pool_contract_address} -nh 127.0.0.1 -np {rpc_port} -ct {network_path}/config/ssl/full_node/private_full_node.crt -ck {network_path}/config/ssl/full_node/private_full_node.key"
     app.logger.info(f"Executing NFT 7/8 win recovery for {blockchain}: {cmd}")
-    log_fo.write("\n\nExecuted at: {0}\nFD_CLI_BC_DB_PATH={1} FD_CLI_WT_DB_PATH={2} {3}".format(
+    log_fo.write("\n\nExecuted at: {0}\nFD_CLI_BC_DB_PATH={1} FD_CLI_WT_DB_PATH={2} {3}\n".format(
         time.strftime("%Y-%m-%d-%H:%M:%S"), vars['FD_CLI_BC_DB_PATH'], vars['FD_CLI_WT_DB_PATH'], cmd))
     log_fo.flush()
     proc = Popen(cmd,cwd="/flora-dev-cli", env=fd_env, shell=True, universal_newlines=True, stdout=log_fo, stderr=log_fo)