+++ common/config/globals.py
@@ -512,3 +512,34 @@ def wallet_running():
     except:
         logging.error(traceback.format_exc())
     return False
+
+def get_container_memory_usage_bytes():
+    try: # Check cgroups v1
+        with open("/sys/fs/cgroup/memory/memory.usage_in_bytes", "r") as f:
+            return int(f.readline())
+    except Exception as ex:
+        pass
+    try: # Check cgroups v2
+        with open("/sys/fs/cgroup/memory.current", "r") as f:
+            return int(f.readline())
+    except Exception as ex:
+        pass
+    return None
+
+def get_host_memory_usage_percent():
+    try:
+        total_mem = None
+        avail_mem = None
+        with open("/proc/meminfo", "r") as f:
+            for line in f.readlines():
+                if line.startswith('MemTotal:'):
+                    #logging.error("{0} -> {1}".format(line, line.split(':')[1].strip().split(' ')[0]))
+                    total_mem = int(line.split(':')[1].strip().split(' ')[0])
+                elif line.startswith('MemAvailable:'):
+                    #logging.error("{0} -> {1}".format(line, line.split(':')[1].strip().split(' ')[0]))
+                    avail_mem = int(line.split(':')[1].strip().split(' ')[0])
+        if total_mem and avail_mem:
+            return 100 - int(avail_mem / total_mem * 100)  # Used memory percentage
+    except Exception as ex:
+        logging.error("Failed to calculate total host memory usage percentage due to: {0}".format(str(ex)))
+    return None