+++ scripts/forks/mmx_install.sh
@@ -4,16 +4,17 @@
 #
 
 MMX_BRANCH=$1
-# On 2022-09-19
-HASH=1d39de976d57aae0b3655d72642bcdf110d4610c
+# On 2022-10-04
+HASH=341442465b1470d8c4b57d1253eaa2e02d4b0b6c
 
 if [ -z ${MMX_BRANCH} ]; then
 	echo 'Skipping MMX install as not requested.'
 else
 	rm -rf /root/.cache
 	apt-get update
-	# Install dependencies
-	apt-get install -y git cmake build-essential libsecp256k1-dev librocksdb-dev libsodium-dev zlib1g-dev ocl-icd-opencl-dev clinfo screen initramfs-tools ocl-icd-libopencl1 opencl-headers apt-utils libnuma1
+	# Install dependencies for MMX and GPU support
+	apt-get install -y git cmake build-essential libsecp256k1-dev libsodium-dev zlib1g-dev ocl-icd-opencl-dev clinfo screen
+	apt-get install -y initramfs-tools ocl-icd-libopencl1 opencl-headers apt-utils libnuma1
 	# For AMDGPU, install the amdgpu-install stub, optionally invoked later if OPENCL_GPU=amd at launch time
 	curl -O http://repo.radeon.com/amdgpu-install/21.40/ubuntu/focal/amdgpu-install-21.40.40500-1_all.deb
 	apt install -y ./amdgpu-install-21.40.40500-1_all.deb