+++ web/actions/stats.py
@@ -26,7 +26,8 @@ from common.models.pools import Pool
 from common.models.partials import Partial
 from common.models.stats import StatPlotCount, StatPlotsSize, StatTotalCoins, StatNetspaceSize, StatTimeToWin, \
         StatPlotsTotalUsed, StatPlotsDiskUsed, StatPlotsDiskFree, StatPlottingTotalUsed, StatEffort, \
-        StatPlottingDiskUsed, StatPlottingDiskFree, StatFarmedBlocks, StatWalletBalances, StatTotalBalance
+        StatPlottingDiskUsed, StatPlottingDiskFree, StatFarmedBlocks, StatWalletBalances, StatTotalBalance, \
+        StatContainerMemoryUsageGib, StatHostMemoryUsagePercent
 from web import app, db, utils
 from web.actions import chia, worker
 
@@ -474,6 +475,46 @@ def load_total_balances(current_currency_symbol):
     return { 'title': _('Wallets Total') + ' (' + current_currency_symbol + ')', 'y_axis_title': _('Fiat Currency'),
          'dates': dates, 'vals': values, 'last_value': " - {0} {1}".format(last_value, current_currency_symbol) }
 
+def load_host_memory_usage():
+    dates = []
+    workers = {}
+    displaynames_for_hosts = {}
+    result = db.session.query(StatHostMemoryUsagePercent).order_by(StatHostMemoryUsagePercent.created_at.asc()).all()
+    for p in result:
+        converted_date = converters.convert_date_for_luxon(p.created_at)
+        if not converted_date in dates:
+            dates.append(converted_date)
+        if p.hostname in displaynames_for_hosts:
+            displayname = displaynames_for_hosts[p.hostname]
+        else:
+            try:
+                w = worker.get_worker(p.hostname)
+                displayname = w.displayname
+            except:
+                app.logger.debug("Failed to find worker for hostname: {0}".format(p.hostname))
+                displayname = p.hostname
+        if not displayname in workers:
+            workers[displayname] = {}
+        values = workers[displayname]
+        values[converted_date] = p.value # Integer as percent of all host memory used
+    values_per_worker = {}
+    if len(dates) > 0:
+        for wk in workers.keys():
+            worker_values = []
+            for date in dates:
+                if wk in workers:
+                    if date in workers[wk]:  
+                        worker_values.append(str(workers[wk][date]))
+                    else: # Due to exeception reported by one user
+                        worker_values.append('null')
+                else:
+                    worker_values.append('null')
+            values_per_worker[wk] = worker_values
+    #app.logger.info(dates)
+    #app.logger.info(values_per_worker)
+    return {'y_axis_title': _('Host Memory Usage') + ' (%)',
+         'dates': dates, "workers": workers.keys(), "values_per_worker": values_per_worker }
+
 def load_netspace_size(blockchain):
     dates = []
     values = []
@@ -503,7 +544,7 @@ def load_farmed_blocks(blockchain):
             w = worker.get_worker(row.hostname)
             displayname = w.displayname
         except:
-            app.logger.debug("Failed to find worker for hostname: {0}".format(ResourceWarning.hostname))
+            app.logger.debug("Failed to find worker for hostname: {0}".format(w.hostname))
             displayname = row.hostname
         blocks.append({
             'hostname': displayname,
@@ -621,6 +662,31 @@ def load_time_to_win(blockchain):
     return { 'title': blockchain.capitalize() + ' - ' + _('ETW'), 'dates': dates, 'vals': converted_values, 
         'y_axis_title': _('Estimated Time to Win') + ' (' + _('days') + ')'}
 
+def load_container_memory(hostname, blockchain):
+    dates = []
+    values = []
+    result = db.session.query(StatContainerMemoryUsageGib).order_by(StatContainerMemoryUsageGib.created_at.asc()).filter(
+            StatContainerMemoryUsageGib.hostname == hostname, StatContainerMemoryUsageGib.blockchain == blockchain).all()
+    for i in range(len(result)):
+        s = result[i]
+        converted_date = converters.convert_date_for_luxon(s.created_at)
+        # First, last, and every 5th (at ~2x5 min intervals)
+        if (i == 0) or (i % 5 == 0) or (i == len(result) - 1):
+            dates.append(converted_date)
+            values.append(s.value)
+    app.logger.debug("{0} before {1}".format(blockchain, values))
+    if len(values) > 0:
+        converted_values = list(map(lambda x: round(x/1024/1024/1024,2), values))  # Bytes to GiB
+    else:
+        converted_values = []
+    app.logger.debug("{0} after {1}".format(blockchain, converted_values))
+    try:
+        displayname = worker.get_worker(hostname).displayname
+    except:
+        displayname = hostname
+    return { 'title': blockchain.capitalize() + ' - ' + _('Container Memory Usage') +  ' - ' + displayname, 'dates': dates, 'vals': converted_values, 
+        'y_axis_title': _('GiB') }
+
 def count_plots_by_type(hostname):
     plots_by_type = {}
     result = db.session.query(Plot.type, func.count(Plot.hostname)).filter(Plot.hostname==hostname).group_by(Plot.type).all()
@@ -656,3 +722,46 @@ def set_disk_usage_per_farmer(farmers, disk_usage):
         else:
             app.logger.info("No disk usage stats found for {0}".format(farmer.hostname))
             farmer.drive_usage = "" # Empty string to report
+
+def load_recent_mem_usage(worker_type, only_hostname=None, only_blockchain=None):
+    summary_by_worker = {}
+    for host in worker.load_workers():
+        hostname = host.hostname
+        if only_hostname and hostname != only_hostname:
+            continue
+        dates = []
+        data_by_blockchain = {}
+        if only_blockchain:
+            mem_result = db.session.query(StatContainerMemoryUsageGib).filter( 
+                StatContainerMemoryUsageGib.hostname == host.hostname, StatContainerMemoryUsageGib.blockchain == only_blockchain). \
+                order_by(StatContainerMemoryUsageGib.created_at, StatContainerMemoryUsageGib.blockchain).all()
+        else: # all blockchains on that hostname
+            mem_result = db.session.query(StatContainerMemoryUsageGib).filter( 
+                StatContainerMemoryUsageGib.hostname == host.hostname). \
+                order_by(StatContainerMemoryUsageGib.created_at, StatContainerMemoryUsageGib.blockchain).all()
+        for row in mem_result:
+            if worker_type == 'plotting' and host.mode != 'plotter' and (host.mode == 'fullnode' and not row.blockchain in ['chia', 'chives', 'mmx']):
+                continue  # Not a plotting container
+            elif worker_type == 'farming' and host.mode == 'plotter':
+                continue # Not a farmer or harvester
+            converted_date = converters.convert_date_for_luxon(row.created_at)
+            if not converted_date in dates:
+                dates.append(converted_date)
+            if not row.blockchain in data_by_blockchain:
+                data_by_blockchain[row.blockchain] = {}
+            data_by_blockchain[row.blockchain][converted_date] = round((row.value / 1024 / 1024 /1024), 2)
+        if len(dates) > 0:
+            summary_by_worker[hostname] = { "dates": dates, "blockchains": data_by_blockchain.keys(),  }
+            for blockchain in data_by_blockchain.keys():
+                blockchain_values = []
+                for date in dates:
+                    if blockchain in data_by_blockchain:
+                        if date in data_by_blockchain[blockchain]:
+                            blockchain_values.append(str(data_by_blockchain[blockchain][date]))
+                        else: 
+                            blockchain_values.append('null')
+                    else:
+                        blockchain_values.append('null')
+                # TODO Decimate the memory usage datapoints as too many being returned...
+                summary_by_worker[hostname][blockchain] = blockchain_values
+    return summary_by_worker